+++ packages/client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tachybase/client",
-  "version": "0.23.8-alpha.27",
+  "version": "0.23.8-alpha.29",
   "license": "Apache-2.0",
   "module": "lib/index.mjs",
   "types": "es/index.d.ts",
