+++ packages/client/src/built-in/pinned-list/PluginPinnedList.ts
@@ -1,0 +1,8 @@
+import { Plugin } from '../../application';
+import { PinnedPluginListProvider } from './PinnedPluginListProvider';
+
+export class PluginPinnedList extends Plugin {
+  async load() {
+    this.app.use(PinnedPluginListProvider, this.options.config);
+  }
+}
