+++ packages/client/src/schema-component/core/SchemaComponentPlugin.tsx
@@ -1,0 +1,14 @@
+import { Plugin } from '../../application';
+import { DesignableSwitch } from './DesignableSwitch';
+
+export class SchemaComponentPlugin extends Plugin {
+  async load() {
+    this.addComponents();
+  }
+
+  addComponents() {
+    this.app.addComponents({
+      DesignableSwitch,
+    });
+  }
+}
